feat(final-adjustments): fixed responsiveness bugs
