+++ src/components/social-network-items/social-network-items.scss
@@ -6,4 +6,7 @@
   &.right {
     right: 0px;
   }
+  @media only screen and (max-width: 1000px) {
+    display: none;
+  }
 }
